* Display exit codes in help message * Exit code for command line error * fixed: abbreviations, exit code on target failure +semver: major

+++ src/Amg.Build/Runner.cs
@@ -34,7 +34,7 @@ namespace Amg.Build
                 rebuildCheck
                 );
 
-            return runner.Run();
+            return (int) runner.Run();
         }
 
         /// <summary>
@@ -76,7 +76,7 @@ namespace Amg.Build
                 rebuildCheck
                 );
 
-            return runner.Run();
+            return (int)runner.Run();
         }
 
         /// <summary>
